Moved plugin source folder
