+++ example/webpack.config.js
@@ -3,7 +3,7 @@ const path = require("path");
 const webpack = require("webpack");
 const ChunkManifestPlugin = require("chunk-manifest-webpack-plugin");
 
-const TemplatedAssetWebpackPlugin = require("../plugins/templated-asset-webpack-plugin");
+const TemplatedAssetWebpackPlugin = require("../templated-assets-webpack-plugin");
 const templatedAssetsConfig = require("./templated-assets-config.js");
 
 module.exports = {
