+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "templated-assets-webpack-plugin",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "webpack plugin for generating templated assets, suitable for i.e. server-side rendering.",
-  "main": "./plugins/templated-asset-webpack-plugin/index.js",
+  "main": "./templated-asset-webpack-plugin/index.js",
   "dependencies": {},
   "devDependencies": {
     "ava": "^0.19.0",
@@ -43,7 +43,7 @@
   },
   "homepage": "https://github.com/jouni-kantola/templated-assets-webpack-plugin#readme",
   "files": [
-    "plugins/templated-asset-webpack-plugin/index.js",
-    "plugins/templated-asset-webpack-plugin/src"
+    "templated-asset-webpack-plugin/index.js",
+    "templated-asset-webpack-plugin/src"
   ]
 }
