+++ templated-assets-webpack-plugin/index.js
