+++ templated-assets-webpack-plugin/src/asset.js
