+++ templated-assets-webpack-plugin/src/compiled-chunks.js
