+++ templated-assets-webpack-plugin/src/logger.js
