+++ templated-assets-webpack-plugin/src/plugin.js
