+++ templated-assets-webpack-plugin/src/rule-set.js
