+++ templated-assets-webpack-plugin/src/template-reader/index.js
