+++ templated-assets-webpack-plugin/src/templated-assets.js
