+++ templated-assets-webpack-plugin/src/templates/async-defer.tmpl
