+++ templated-assets-webpack-plugin/src/templates/async.tmpl
