+++ templated-assets-webpack-plugin/src/templates/defer.tmpl
