+++ templated-assets-webpack-plugin/src/templates/inline.tmpl
