+++ templated-assets-webpack-plugin/src/templates/sync.tmpl
