+++ templated-assets-webpack-plugin/test/asset-file-test.js
