+++ templated-assets-webpack-plugin/test/asset-init-test.js
