+++ templated-assets-webpack-plugin/test/asset-template-replace-test.js
