+++ templated-assets-webpack-plugin/test/asset-template-test.js
