+++ templated-assets-webpack-plugin/test/asset-type-test.js
