+++ templated-assets-webpack-plugin/test/compiled-chunks-init-test.js
