+++ templated-assets-webpack-plugin/test/rule-set-filter-test.js
