+++ templated-assets-webpack-plugin/test/rule-set-init-test.js
