+++ templated-assets-webpack-plugin/test/template-reader-test.js
