+++ templated-assets-webpack-plugin/test/templated-assets-init-test.js
